# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/advertisers": {
      "post": {
        "operationId": "AdvertisersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdvertiserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      },
      "get": {
        "operationId": "AdvertisersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      }
    },
    "/advertisers/register": {
      "post": {
        "operationId": "AdvertisersController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdvertiserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      }
    },
    "/advertisers/{id}": {
      "get": {
        "operationId": "AdvertisersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      },
      "patch": {
        "operationId": "AdvertisersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdvertiserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      },
      "delete": {
        "operationId": "AdvertisersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Advertisers"
        ]
      }
    },
    "/campaigns": {
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/expiring": {
      "get": {
        "operationId": "CampaignsController_getExpiringCampaigns",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      },
      "delete": {
        "operationId": "CampaignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_modifyCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}/activate": {
      "post": {
        "operationId": "CampaignsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}/cancel": {
      "post": {
        "operationId": "CampaignsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}/complete": {
      "post": {
        "operationId": "CampaignsController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/campaigns/{id}/update-creative": {
      "post": {
        "operationId": "CampaignsController_updateCreative",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCreativeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Campaigns"
        ]
      }
    },
...
    "/reports/campaign/{campaignId}": {
      "get": {
        "operationId": "ReportsController_getCampaignReport",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "today",
                "week",
                "month",
                "quarter",
                "lifetime"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "AdOnTheGo API",
    "description": "AdOnTheGo Backend APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Enter: ApiKey <your-api-key>"
      }
    },
    "schemas": {
      "CreateAdvertiserDto": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "example": "AdOnTheGo"
          },
          "contactName": {
            "type": "string",
            "example": "Hemil Kothari"
          }
        },
        "required": [
          "companyName",
          "contactName"
        ]
      },
      "UpdateAdvertiserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "campaignName": {
            "type": "string",
            "example": "ABC Campaign"
          },
          "advertiserId": {
            "type": "string",
            "example": "adv_123456"
          },
          "durationDays": {
            "type": "number",
            "example": 30
          },
          "frequencyPerLoop": {
            "type": "number",
            "example": 5
          },
          "fleetSize": {
            "type": "number",
            "example": 1000
          },
          "mediaAssetId": {
            "type": "string",
            "description": "Uploaded Media Asset ID"
          },
          "creativeChangesAllowed": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "campaignName",
          "advertiserId",
          "durationDays",
          "frequencyPerLoop",
          "fleetSize",
          "mediaAssetId",
          "creativeChangesAllowed"
        ]
      },
      "UpdateCreativeDto": {
        "type": "object",
        "properties": {}
      },
      "ModifyCampaignDto": {
        "type": "object",
        "properties": {}
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}